Task 4.2 //Подключил приложение к REST API IMDB - теперь грузим два списка фильмов с сервера
** рефакторинг кода (удалены неиспользуемые ресурсы, импорты и переименованы функции в соответствии с тем, что они теперь выполняют)
