+++ app/build.gradle
@@ -49,5 +49,6 @@ dependencies {
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+    implementation 'com.google.code.gson:gson:2.8.9'
     implementation 'pl.droidsonroids.gif:android-gif-drawable:1.2.24'
 }
