--- app/src/main/java/ru/kinomafia/model/Repository.kt
@@ -1,7 +1,0 @@
-package ru.kinomafia.model
-
-interface Repository {
-    fun getFilmInfoFromServer(): FilmItem
-    fun getFilmInfoFromLocalStorageNovelties(): List<FilmItem>
-    fun getFilmInfoFromLocalStorageHits(): List<FilmItem>
-}
