--- app/src/main/java/ru/kinomafia/model/RepositoryImpl.kt
@@ -1,7 +1,0 @@
-package ru.kinomafia.model
-
-class RepositoryImpl: Repository {
-    override fun getFilmInfoFromServer() = getDefaultFilm()
-    override fun getFilmInfoFromLocalStorageNovelties() = getNoveltiesFilms()
-    override fun getFilmInfoFromLocalStorageHits() = getHitsFilms()
-    }
