+++ app/src/main/java/ru/kinomafia/model/entities/rest_entities/FilmItemListDTO.kt
@@ -1,0 +1,5 @@
+package ru.kinomafia.model.entities.rest_entities
+
+data class FilmItemListDTO (
+    val items: List<ItemsDTO>
+)
