+++ app/src/main/java/ru/kinomafia/model/entities/rest_entities/ItemsDTO.kt
@@ -1,0 +1,9 @@
+package ru.kinomafia.model.entities.rest_entities
+
+data class ItemsDTO(
+    val id: String?,
+    val title:String?,
+    val year: String?,
+    val image: String?,
+    val imDbRating: String?
+)
