+++ app/src/main/java/ru/kinomafia/model/repository/Repository.kt
@@ -1,0 +1,8 @@
+package ru.kinomafia.model.repository
+
+import ru.kinomafia.model.entities.FilmItem
+
+interface Repository {
+    fun getMostPopularMoviesFilmListFromServer(): List<FilmItem>
+    fun getTop250FilmListFromServer(): List<FilmItem>
+}
