+++ app/src/main/java/ru/kinomafia/viewmodel/AppState.kt
@@ -1,6 +1,6 @@
 package ru.kinomafia.viewmodel
 
-import ru.kinomafia.model.FilmItem
+import ru.kinomafia.model.entities.FilmItem
 
 sealed class AppState {
     data class Success(val filmDataNovelties: List<FilmItem>, val filmDataHits: List<FilmItem>): AppState()
