+++ app/src/main/res/layout/film_info_fragment.xml
@@ -11,14 +11,14 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         xmlns:app="http://schemas.android.com/apk/res-auto"
-        tools:context=".view.FilmInfoFragment">
+        tools:context=".view.film_info.FilmInfoFragment">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/main"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginTop="@dimen/margin_film_info_fragment"
-            tools:context=".view.FilmInfoFragment">
+            tools:context=".view.film_info.FilmInfoFragment">
 
             <TextView
                 android:id="@+id/name_film_info"
@@ -43,7 +43,7 @@
                 app:layout_constraintEnd_toEndOf="parent"
                 android:adjustViewBounds="true"
                 android:layout_marginEnd="@dimen/margin_film_info_fragment"
-                tools:src="@drawable/terminator2poster"/>
+                tools:src="@drawable/poster_no"/>
 
             <TextView
                 android:id="@+id/genre_label"
