+++ app/src/main/res/layout/main_fragment.xml
@@ -23,7 +23,7 @@
     android:layout_marginBottom="@dimen/margin_film_info_fragment"
     android:layout_marginStart="@dimen/margin_film_info_fragment"
     android:layout_marginEnd="@dimen/margin_film_info_fragment"
-    tools:context=".view.MainFragment">
+    tools:context=".view.main.MainFragment">
 
     <TextView
         android:id="@+id/hits_title_main_fragment"
@@ -31,7 +31,7 @@
         android:textStyle="bold"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="@string/hits_title_main_fragment" />
+        android:text="@string/top250_title_main_fragment" />
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/hits_recycler_main_fragment"
@@ -48,7 +48,7 @@
         android:textStyle="bold"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="@string/novelties_title_main_fragment" />
+        android:text="@string/most_popular_title_main_fragment" />
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/novelties_recycler_main_fragment"
