+++ app/src/test/java/ru/kinomafia/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
 package ru.kinomafia
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
